+++ src/main/java/org/joelson/turf/scorecalc/imprt/model/AssistImportRepository.java
@@ -10,8 +10,8 @@ import java.util.Optional;
 public interface AssistImportRepository extends JpaRepository<AssistImport, AssistImportId> {
 
     @Query("select a from AssistImport a where a.zone.id = :zoneId and a.time = :time and a.user.id = :userId")
-    <T> Optional<T> findByZoneIdAndTimeAndUserId(Long zoneId, Instant time, Long userId, Class<T> type);
+    <T> Optional<T> findByZoneIdAndTimeAndUserId(Integer zoneId, Instant time, Integer userId, Class<T> type);
 
     @Query("select a from AssistImport a where a.zone.id = :zoneId and a.time = :time order by a.user.id")
-    <T> List<T> findByZoneIdAndTime(Long zoneId, Instant time, Class<T> type);
+    <T> List<T> findByZoneIdAndTime(Integer zoneId, Instant time, Class<T> type);
 }
