+++ src/main/java/org/joelson/turf/scorecalc/imprt/model/RegionImportRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RegionImportRepository extends JpaRepository<RegionImport, Long> {
+public interface RegionImportRepository extends JpaRepository<RegionImport, Integer> {
 
-    <T> Optional<T> findByRegionId(Long regionId, Class<T> type);
+    <T> Optional<T> findByRegionId(Integer regionId, Class<T> type);
 }
