+++ src/main/java/org/joelson/turf/scorecalc/imprt/model/UserImportRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserImportRepository extends JpaRepository<UserImport, Long> {
+public interface UserImportRepository extends JpaRepository<UserImport, Integer> {
 
-    <T> Optional<T> findByUserId(Long userId, Class<T> type);
+    <T> Optional<T> findByUserId(Integer userId, Class<T> type);
 }
