+++ src/main/java/org/joelson/turf/scorecalc/imprt/model/VisitImportRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 public interface VisitImportRepository extends JpaRepository<VisitImport, VisitImportId> {
 
     @Query("select v from VisitImport v where v.zone.id = :zoneId and v.time = :time")
-    <T> Optional<T> findByZoneIdAndTime(Long zoneId, Instant time, Class<T> type);
+    <T> Optional<T> findByZoneIdAndTime(Integer zoneId, Instant time, Class<T> type);
 }
